Realplayer Browser Fix
The REALPLAYER tests for the presence of the RealPlayer plugin or the
browser. This has led to MSIE, with Realplayer installed to be labeled
as Realplayer browser. 

Added a new subroutine realplayer_browser that returns true only if the
realplayer test detected "RealPlayer" in the UA and called the
_realplayer_version method. The browser_string method now calls
realplayer_browser instead of the REALPLAYER test.

The addition of the new sub preserves the functionality of the
existing test and corrects the flaw in browser_string.