RealPlayer Browser vs. plugin fix
The original REALPLAYER test looks for the presence of either the
plug-in or the browser. In order to preserve backward compatibility 
I added a new sub realplayer_browser that browser_string() uses to
identify the user agent as a RealPlayer browser. This should prevent
MSIEs with realplayer installed from being falsely identified as
Realplayer browsers.